feat: tabs & team numbers
