--- .env.example
@@ -1,2 +1,0 @@
-VITE_TENCENT_SMS_SECRET_ID=your-secret-id
-VITE_TENCENT_SMS_SECRET_KEY=your-secret-key
