+++ src/components/form.json
@@ -1,0 +1,127 @@
+[
+    {
+      "question": "",
+      "type": "hidden",
+      "required": true,
+      "value": "_event"
+    },
+    {
+      "question": "Team number",
+      "type": "select",
+      "required": true,
+      "value": null
+    },
+    {
+      "i": "https://lh7-us.googleusercontent.com/pUWvHrPDa5IfrQcFalk4lO0e4PhD3sLMP0jyLJU8PTWWGfw5r-Wa4qDQNHhbu0byYLzXScP5lfTSUCsvbNI-FlwDY2L7Ra0-TgYqf5Eabw0INSFE3ah4QCqCqHFrsaPKyCOt8m2Yo-H2ie9E7apzh6c8AO147A",
+      "w": "50%",
+      "question": "Type of drive train",
+      "type": "radio",
+      "options": [
+        "Tank Drive ('skid steer', plates on both sides of wheels)",
+        "West Coast Drive (wheels mounted off one side of tube)",
+        "Swerve Drive"
+      ],
+      "value": null,
+      "required": true,
+      "showOtherInput": false, 
+      "otherValue": ""
+    },
+    {
+      "i": "https://lh7-us.googleusercontent.com/PCI7CaG88MiY50L7AM0CVTs9dRd3NQgqW4B2rd64vmjHaNDMEHR0EkWYqv-rzHBnGBC08NzWtr7W97lIk226Q9WVCPuTKuOSZcpb6eyNC5Q3HGmFQwp8005gRcxiS09RjeWUJQJTK-vQGDWd0QAbpSipLSkExw",
+      "w": "100%",
+      "question": "Type of wheels used",
+      "type": "radio",
+      "options": [
+        "Traction",
+        "Mecanum (rollers at 45° angle)",
+        "Omni (rollers at 90° angle)"
+      ],
+      "value": null,
+      "required": true,
+      "showOtherInput": false,
+      "otherValue": ""
+    },
+    {
+      "question": "Intake Use:",
+      "type": "checkbox",
+      "options": ["Ground", "Station"],
+      "value": [], 
+      "required": true,
+      "showOtherInput": false, 
+      "otherValue": ""
+    },
+    {
+      "question": "Scoring Locations:",
+      "type": "checkbox",
+      "options": ["Amp", "Speaker", "Trap", "Hang", "Harmony"],
+      "value": [], 
+      "required": true,
+      "showOtherInput": false, 
+      "otherValue": ""
+    },
+    {
+      "question": "Robot Weight (in pounds)",
+      "type": "number",
+      "required": true,
+      "value": null
+    },
+    {
+      "question":
+        "Robot Dimension (Length in Inches) without bumpers - front to back",
+      "type": "number",
+      "required": true,
+      "value": null
+    },
+    {
+      "question":
+        "Robot Dimension (Width in Inches) without bumpers - left to right",
+      "type": "number",
+      "required": true,
+      "value": null
+    },
+    {
+      "question":
+        "Robot Dimension (Height in Inches) from floor to highest point on robot at the start of the match",
+      "type": "number",
+      "required": true,
+      "value": null
+    },
+    {
+      "question": "Drive Team Members",
+      "type": "radio",
+      "options": [
+        "One person driving and operating the robot during a match"
+      ],
+      "value": null,
+      "required": true,
+      "showOtherInput": false
+    },
+
+    {
+      "question": "Maneuverability",
+      "type": "checkbox",
+      "options": ["Can it drive under the core"],
+      "value": [], 
+      "required": false,
+      "showOtherInput": false,
+      "otherValue": ""
+    },
+    {
+      "question": "Height when fully extended (in inches)",
+      "type": "number",
+      "required": true,
+      "value": null
+    },
+    {
+      "question": "Hours/Weeks of Practice",
+      "type": "text",
+      "required": true,
+      "value": null
+    },
+    {
+      "question": "Additional Comments",
+      "type": "textarea",
+      "required": false,
+      "value": null
+    }
+  ]
