+++ vite.config.js
@@ -6,8 +6,7 @@ import { ElementPlusResolver } from 'unplugin-vue-components/resolvers'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-	publicPath: './',
-	base: "/",
+	base: "/pit-scouting/",
 	server: {
 		port: 8080
 	},
